Get nightly builds working
There were several changes needed to get SSL and nightly builds working.
As part of this change, we have added a CI image specificly for building
the nightly releases as well as a CI job to keep it updated.

Because we build the nightlies in alpine, we have to use SSL 0.9.0
as it is what the libressl there supports.