Make default SSL version match our test and build environments
We use libressl, which matches the openssl 0.9.0 API, in all our
CI environments.

This change switches the default in our Makefile from 1.1.x to 0.9.0
thereby making maintenance of all our scripts much easier. By doing this,
all the release scripts for ponyup will be the same as changelog-tool and
pony-stable.

That is going to make maintenance on the scripts much easier as we can
copy changes around.

This is a prelude to issue #35 and will make #35 much less error prone.