Detect platform identifier with init script (#60)
Detect platform using a `.platform` config file created by the init script. This also cleans up much of the code for the `show` command and adds tests for the platform string parsing.

With this change, supplying `--platform=musl` on musl-based Linux distros like Alpine should no longer be required as "musl-based" should be detected when ponyup is installed. 

If ponyup is upgraded from an earlier version without using the init script, then this change will have no impact and `--platform=musl` will continue to be needed.