Fix .platform creation
Previously, we were writing out the contents of `cc -dumpmachine`
directly to the .platform file which could be incorrect for what
we expect.

Given that we have already determined the CPU and OS parts of our
triple, we can use those and only use `cc -dumpmachine` on Linux
to determine the libc version- be it glibc or musl.