Allow for ponyup-init.sh to only change on releases
Currently, any change on master is immediately reflected in a
change for end users. This happens because we have them download
from master. This can and has lead to bugs being exposed
to users.

With this change, we will instead have a tag that gets moved on each
release that will instead be used. Having users download the "latest-release"
tagged version of ponyup-init.sh will allow us to protect users from
random breakage.

After this is approved but before it is merged, we need to create an
initial "latest-release" tag so that the new url in README.md will work
prior to the next release.