Add nightly macOS breakage tests
We PR against macOS as well as Linux so it makes sense to run
the nightly breakage test against on both Linux and macOS as well.

Note, because macOS ponyc nightly is currently broken, this will
start failing once it's merged. Once ponyc nightly is fixed, this
should start passing.