Update bootstrap environments
We need clang to be installed with most recent ponyc as it uses
clang as the default linker.

Also, update to Ubuntu 20.04 as we no longer support 18.04

The Fedora environment is broken because our glibc builds don't work on it.
This PR removes bootstrap testing on Fedora; we will readd in the future if
Fedora is a supported platform.