improve regex match of json field in init script
Modifications to the Cloudsmith API have resulted in mangled version identifiers for the initial ponyup install. This has resulted in the `"tags": {"version": ["latest"]},` being matched as the version, resulting in the following output for `ponyup show`:
```
ponyup-release-[latest]}-x86_64-linux
```
The updated regex should be sufficient to prevent such an issue from occurring in the future without requiring `jq` to be available on the user's system.