Fixes Issue #191
The root of this bug is where the expansion of \e to the
binary escape character happens.  In the case before this PR
it's being done by printf.

FreeBSD does not have \e in its printf implementation.
Linux does.

The fix I am suggesting is to get the shell to do the expansion
instead.  By changing the string definitions from using double
quotes to dollar single-quote format - the shell will do the
expansion.