Never use the "generic gnu" ponyc package
The "generic gnu" package is not actually generic. It is a glibc
version that is built on "some glibc using Linux" and will work if
it is installed on a Linux that is compatible at a library level with
the system that we built the "generic gnu" package on.

Quite some time ago, Theo and I (who made the initial decision to include
a "generic gnu") realized that the "generic gnu" idea was "a very bad
idea".

There are a couple large problems with this approach:

1- anyone using ponyup on glibc distro that isn't compatible with our
"generic gnu" platform will by default have things installed that will
not work. That's a bad user experience.

2- in order to not fall horribly behind, from time to time, we need to
"rev" the environment used to build "generic gnu" in. For us that has
meant changing the Ubuntu version periodically. That then means that
people who had an environment that worked will do an update and install
a ponyc that doesn't work because it is for a different set of libraries.
That's a bad user experience as well.

This PR drops "generic gnu" from being used by the ponyup installer and
it avoids it being set as a target platform. Instead when the installer
detects that it is being run on a glibc based Linux, it will attempt to
identify the distro being used and if we support it, it will use the
builds we do for that distro. In the case of "distros built on distros"
like Linux Mint being built on Ubuntu, it will set the platform to the
builds for the appropriate "base distro". If we are unable to determine
the distro or the distro isn't supported, an error message is displayed
that directs them to the Linux install instructions in the ponyc repo.

I will be doing a corresponding PR to ponyc to update the Linux section
of INSTALL.md to account for the changes from this PR. Additionally, once
this is merged, we can stop building the "generic gnu" package which I will
also open a PR for.