Fix race condition in sync tests
The nightly and release tests could step on each other. Both could
be running at the same time and could be executing the symlink update
in an interleaved fashion.

We now do nightly and release sync tests in different asset directories
to have them avoid stepping on one another.

Fixes #279